train_ml/requirements.txt
# Data Collection Orchestrator Requirements

# ==============================================
# Core Data Processing
# ==============================================
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2

# ==============================================
# Database
# ==============================================
pymongo>=4.6.0
motor>=3.3.0

# ==============================================
# Task Queue & Scheduling
# ==============================================
redis>=5.0.0
celery>=5.3.0
flower>=2.0.0  # Celery monitoring

# ==============================================
# Web Framework (Monitoring Dashboard)
# ==============================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# ==============================================
# Web Scraping
# ==============================================
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
firecrawl-py>=0.0.8
praw>=7.7.1  # Reddit scraping
aiohttp>=3.9.0

# ==============================================
# AI & NLP
# ==============================================
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.5

# ==============================================
# Data Cleaning & Validation
# ==============================================
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

# ==============================================
# Machine Learning
# ==============================================
scikit-learn>=1.3.0
scipy>=1.11.0

# ==============================================
# Visualization
# ==============================================
matplotlib>=3.7.0
seaborn>=0.12.0

# ==============================================
# Configuration & CLI
# ==============================================
pyyaml>=6.0.1
rich>=13.7.0
click>=8.1.7
python-dotenv>=1.0.0
pydantic>=2.5.0
tenacity>=8.2.3

# ==============================================
# Email Notifications (Optional)
# ==============================================
# sendgrid>=6.10.0
# smtplib is built-in

# ==============================================
# Testing (Optional)
# ==============================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# ==============================================
# Type Checking (Optional)
# ==============================================
# mypy>=1.5.0
# types-pyyaml
# types-redis
# types-requests
